Fixed fe products

+++ backend/database/migrations/2023_11_19_152601_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('description');
             $table->float('price');
-            $table->integer('star_avg')->default(0);
+            $table->float('star_avg')->default(0);
             $table->timestamps();
         });
     }
